default layout changed

+++ app/views/admin_users/show.blade.php
@@ -1,4 +1,5 @@
-@extends('layouts.scaffold')
+@extends('layouts.default')
+@section('title', 'Admin')
 
 @section('main')
 
@@ -22,12 +23,14 @@
 					<td>{{{ $admin_user->email }}}</td>
 					<td>{{{ $admin_user->password }}}</td>
 					<td>{{{ $admin_user->is_superuser }}}</td>
-                    <td>
-                        {{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin_users.destroy', $admin_user->id))) }}
-                            {{ Form::submit('Delete', array('class' => 'btn btn-danger')) }}
-                        {{ Form::close() }}
-                        {{ link_to_route('admin_users.edit', 'Edit', array($admin_user->id), array('class' => 'btn btn-info')) }}
-                    </td>
+                    @if (Auth::user()->is_superuser)
+                        <td>
+                            {{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin_users.destroy', $admin_user->id), 'onsubmit' => 'return ConfirmDelete()')) }}
+                                {{ Form::submit('Delete', array('class' => 'btn btn-danger')) }}
+                            {{ Form::close() }}
+                            {{ link_to_route('admin_users.edit', 'Edit', array($admin_user->id), array('class' => 'btn btn-info')) }}
+                        </td>
+                    @endif
 		</tr>
 	</tbody>
 </table>
