+++ app/views/documents/show.blade.php
@@ -1,4 +1,5 @@
-@extends('layouts.scaffold')
+@extends('layouts.default')
+@section('title', 'Document')
 
 @section('main')
 
@@ -25,47 +26,12 @@
 
     @if (Auth::user()->is_superuser)
         <div class="pos_right">
-            {{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin.documents.destroy', $document->id))) }}
+            {{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin.documents.destroy', $document->id), 'onsubmit' => 'return ConfirmDelete()')) }}
             {{ Form::submit('Delete', array('class' => 'btn btn-danger')) }}
             {{ Form::close() }}
 
         {{ link_to_route('admin.documents.edit', 'Edit', array($document->id), array('class' => 'btn btn-info')) }}
         </div>
     @endif
-
-
-
-{{--<table class="table table-striped">--}}
-	{{--<thead>--}}
-		{{--<tr>--}}
-			{{--<th>Url</th>--}}
-            {{--<th>Input_format</th>--}}
-            {{--<th>Output_format</th>--}}
-            {{--<th>Api_version</th>--}}
-            {{--<th>Description</th>--}}
-            {{--<th>Request_method</th>--}}
-		{{--</tr>--}}
-	{{--</thead>--}}
-
-	{{--<tbody>--}}
-		{{--<tr>--}}
-			{{--<td>{{{ $document->url }}}</td>--}}
-					{{--<td><pre>{{{ $document->input_format }}}</pre></td>--}}
-					{{--<td><pre>{{{ $document->output_format }}}</pre></td>--}}
-					{{--<td>{{{ $document->api_version }}}</td>--}}
-					{{--<td>{{{ $document->description }}}</td>--}}
-					{{--<td>{{{ $document->request_method }}}</td>--}}
-                    {{--<td>--}}
-                        {{--{{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin.documents.destroy', $document->id))) }}--}}
-                            {{--{{ Form::submit('Delete', array('class' => 'btn btn-danger')) }}--}}
-                        {{--{{ Form::close() }}--}}
-
-                    {{--</td>--}}
-                    {{--<td>--}}
-                        {{--{{ link_to_route('admin.documents.edit', 'Edit', array($document->id), array('class' => 'btn btn-info')) }}--}}
-                    {{--</td>--}}
-		{{--</tr>--}}
-	{{--</tbody>--}}
-{{--</table>--}}
 
 @stop
